requirements.txt
python-dotenv==1.0.1
psycopg[binary]==3.2.3
fastapi==0.115.0
uvicorn==0.31.0
PyMuPDF==1.24.11
azure-identity==1.19.0
azure-storage-blob==12.23.1
cachetools==5.5.0
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-openai==0.2.3
SQLAlchemy==2.0.35